Suggestion for rephrasing.
I think there is a little mistake when talking
about the percentage of children havin some number
of symptom-free days. There is not 50% having 3
symptom-free days, but 50% have more symptom-free
days, and 50% percent have less than 3 symptom-free
days. It is impossible to tell if 50% of the children
have 3 symptom-free days according to me.