normalizePath is the right command to return full path
According to the definition of Unix paths[1][2], "full path" refers to
the absolute path, unique location regardless of the present working
directory.

> file.path("mytest3.R")
[1] "mytest3.R"
> normalizePath("mytest3.R")
[1] "/Users/irio/Code/r_programming/assignment_1/testdir/mytest3.R"

---

[1]:
"Systems can use either absolute or relative paths. A full path or
absolute path is a path that points to the same location on one file
system regardless of the present working directory or combined paths. As
such it must always contain the root directory."
http://en.wikipedia.org/wiki/Path_%28computing%29

[2]:
"A full pathname is just a path that starts at the root directory, i.e.,
begins with the slash (/) character.
Any example full path is:
    /usr/bin/man
which is the location of the man command."
https://www.cs.bu.edu/teaching/unix/reference/vocab.html