changing the hint at 19%
The number of odd numbers between 1 and 6 is 3. The reciprocal of 3 is 1/3, but the answer is 1/2. I would like to suggest the wording should be either "The probability of B is the reciprocal of the fraction of odd numbers to the number of sides of a die" or "The probability of B is the probability of an odd number occurring".