clarifying the transition between statements
I was unable to understand the jump between the above statements. I posted the question in the related Coursera forum and Alan Berger (mentor) was kind enough to break it down for me. Using Alan's explanation, I offer this suggestion to further enhance the swirl lesson between the statements at 26% and 29%.